% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poison_frog_palettes.R
\docType{data}
\name{poison_palettes}
\alias{poison_palettes}
\title{Poison frog palettes}
\format{
An object of class \code{list} of length 42.
}
\usage{
poison_palettes
}
\description{
Poison frog palettes
}
\keyword{internal}
