% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poison_frog_palettes.R
\name{poison_palettes_names}
\alias{poison_palettes_names}
\title{List the names of available color palettes inspired in poison frogs and other frog species.}
\usage{
poison_palettes_names()
}
\value{
A character vector of palette names.
}
\description{
List the names of available color palettes inspired in poison frogs and other frog species.
}
\examples{

poison_palettes_names()

}
