% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisonfrogs-package.R
\docType{package}
\name{poisonfrogs}
\alias{poisonfrogs-package}
\alias{poisonfrogs}
\title{poisonfrogs: Colour Palettes Inspired by Neotropical Poison Frogs}
\description{
Provides colour palettes inspired by the enormous diversity of skin colours in Neotropical poison frog species. Suitable for use with ggplot2 and base R graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://laurenoconnelllab.github.io/poisonfrogs/}
}

}
\author{
\strong{Maintainer}: Camilo Rodríguez \email{camilorl@stanford.edu}

Authors:
\itemize{
  \item Lauren O'Connell \email{loconnel@stanford.edu}
}

}
