% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poison_frog_palettes.R
\name{show_poison}
\alias{show_poison}
\title{Show a poison frog palette as coloured tiles}
\usage{
show_poison(
  name,
  n = NULL,
  type = c("discrete", "continuous"),
  direction = 1,
  alpha = NULL
)
}
\arguments{
\item{name}{Palette name (see \code{\link[=poison_palettes_names]{poison_palettes_names()}}).}

\item{n}{Number of colours to show (defaults to full length for discrete,
or 16 for continuous).}

\item{type}{"discrete" or "continuous".}

\item{direction}{\code{1} forward, \code{-1} reversed.}

\item{alpha}{Optional transparency in \link{0, 1}.}
}
\value{
A ggplot2 object (plot of the palette).
}
\description{
Show a poison frog palette as coloured tiles
}
