% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poison_frog_palettes.R
\name{poison_palettes_names}
\alias{poison_palettes_names}
\title{List available poison palettes}
\usage{
poison_palettes_names()
}
\value{
A character vector of palette names.
}
\description{
List available poison palettes
}
\examples{

poison_palettes_names()

}
