% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisonfrogs-package.R
\docType{package}
\name{poisonfrogs}
\alias{poisonfrogs-package}
\alias{poisonfrogs}
\title{Colour Palettes of Neotropical Poison Frogs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A collection of color palettes inspired by the enormous diversity of skin colors in Neotropical poison frog species. Suitable for use with ggplot2 and base R graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://laurenoconnelllab.github.io/poisonfrogs/}
  \item Report bugs at \url{https://github.com/laurenoconnelllab/poisonfrogs/issues}
}

}
\author{
\strong{Maintainer}: Camilo Rodríguez \email{camilorl@stanford.edu}

Authors:
\itemize{
  \item Lauren O'Connell \email{loconnel@stanford.edu}
}

}
